requirements.txt
Django==4.1.3
djangorestframework==3.14.0
Markdown==3.4.1
python-dotenv==0.21.0
psycopg2-binary==2.9.5
django-filter==22.1